pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["edge_catcher*"]

[project]
name = "edge-catcher"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31",
    "pandas>=2.0",
    "scipy>=1.11",
    "numpy>=1.24",
    "statsmodels>=0.14",
    "pyyaml>=6.0",
    "psutil>=6.0",
]

[project.optional-dependencies]
live = [
    "websockets>=12.0",
    "aiohttp>=3.9",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
ai = [
    "anthropic>=0.40",
    "openai>=1.0",
    "python-dotenv>=1.0",
]
ui = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-multipart>=0.0.9",
]

[project.scripts]
edge-catcher = "edge_catcher.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
